+++ CHANGELOG.md
@@ -6,7 +6,18 @@
 * Create methods for Sample and Project
     * .describe [ ]
     * .write_paths_to_file, .copy_files_to_dir(), .link_files_to_dir() [ ]
+    * .total_duration [ ]
 * Add logger calls when saving to JSON and uploading to ProvStore [ ]
+
+### v0.1.23
+* Patch PresetProgram SeqFile addition feature [x]
+* Fix SeqFile deserializer [x]
+* Add import_records arg to `bp.load_project()` [x]
+* Improve reserved aminoacid characters [x]
+* Add `SeqFile.max_seq` and `.min_seq` properties [x]
+* Patch `Project` deserializer to improve BioProvDocument creation [x]
+* Make shorter Environment hashes [x]
+* Improve Project `__repr__` [x]
 
 ### v0.1.22
 * Simplify `bp.load_project()` function [x]